requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
delocate==0.12.0
docopt==0.6.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
macholib==1.16.3
moviepy==1.0.3
numpy==2.1.2
packaging==24.1
pillow==11.0.0
pipreqs==0.4.13
proglog==0.1.10
requests==2.32.3
scipy==1.14.1
setuptools==75.2.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yarg==0.1.10
